+++ pkgs/client/templates/jig.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://deploywithjig.askh.at/jsonschema.json",
+  "name": "dir-name"
+}
