+++ pkgs/client/templates/node.Dockerfile
@@ -1,0 +1,2 @@
+FROM node:latest
+
